Correction Main.class

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int CurrentBalance = 2_000_000_000;
-        int MoneyTransfer = 500_000_000;
-        int EndBalance = CurrentBalance + MoneyTransfer;
+        long CurrentBalance = 2_000_000_000;
+        long MoneyTransfer = 500_000_000;
+        long EndBalance = CurrentBalance + MoneyTransfer;
         System.out.println("EndBalance =" + EndBalance);
     }
 }
